// src/infrastructure/database/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const CreditType = {
  BONUS: 'BONUS',
  USAGE: 'USAGE',
  PURCHASE: 'PURCHASE',
  REFUND: 'REFUND'
} as const

export type CreditType = (typeof CreditType)[keyof typeof CreditType]


export const FileType = {
  INPUT_AUDIO: 'INPUT_AUDIO',
  OUTPUT_AUDIO: 'OUTPUT_AUDIO'
} as const

export type FileType = (typeof FileType)[keyof typeof FileType]


export const JobStatus = {
  QUEUED: 'QUEUED',
  PROCESSING: 'PROCESSING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED'
} as const

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus]